organize all shuffles in one module
