+++ src/cloak.rs
@@ -1,5 +1,6 @@
 use bulletproofs::r1cs::{ConstraintSystem, R1CSError};
-use crate::{merge, padded_shuffle, range_proof, split, value_shuffle};
+use crate::{merge, range_proof, split};
+use shuffle::{padded_shuffle, value_shuffle};
 use value::AllocatedValue;
 
 /// Enforces that the outputs are a valid rearrangement of the inputs, following the
@@ -19,17 +20,17 @@ pub fn cloak<CS: ConstraintSystem>(
     // Shuffle 1
     // Check that `merge_in` is a valid reordering of `inputs`
     // when `inputs` are grouped by flavor.
-    value_shuffle::fill_cs(cs, inputs, merge_in)?;
+    value_shuffle(cs, inputs, merge_in)?;
 
     // Shuffle 2
     // Check that `split_in` is a valid reordering of `merge_out`, allowing for
     // the adding or dropping of padding values (quantity = 0) if m != n.
-    padded_shuffle::fill_cs(cs, merge_out, split_in)?;
+    padded_shuffle(cs, merge_out, split_in)?;
 
     // Shuffle 3
     // Check that `split_out` is a valid reordering of `outputs`
     // when `outputs` are grouped by flavor.
-    value_shuffle::fill_cs(cs, split_out, outputs.clone())?;
+    value_shuffle(cs, split_out, outputs.clone())?;
 
     // Range Proof
     // Check that each of the quantities in `outputs` lies in [0, 2^64).
