+++ src/lib.rs
@@ -1,4 +1,5 @@
 extern crate bulletproofs;
+extern crate core;
 extern crate curve25519_dalek;
 extern crate merlin;
 extern crate rand;
@@ -9,19 +10,16 @@ mod merge;
 mod mix;
 mod split;
 
-mod padded_shuffle;
-mod scalar_shuffle;
-mod value_shuffle;
+mod shuffle;
 
-mod range_proof;
 mod cloak;
+mod range_proof;
 
 mod value;
 
 pub use cloak::cloak;
 pub use range_proof::range_proof;
-pub use value::{Value, CommittedValue};
+pub use value::{CommittedValue, Value};
 
 // TBD: figure out if we need to export these at all
 pub use value::{ProverCommittable, VerifierCommittable};
-
